+++ src/utils/events.ts
@@ -6,8 +6,6 @@
  * @author: Sebastian Fitzner
  */
 
-
-
  /**
  * Events Global
  */