+++ src/js/utils/helpers/is-touch.js
@@ -1,8 +1,0 @@
-'use strict';
-
-/**
- * Touch Detection
- */
-export default function isTouch() {
-	return 'ontouchstart' in window;
-};